Rename namespace to `ORM`. Update docs.
